added correct test.py to follow instructions
